refactor: component split and state management service
